+++ Examples/UIKit-Swift/Superwall-UIKit-Swift/HomeViewController.swift
@@ -10,19 +10,19 @@ import UIKit
 import SuperwallKit
 import Combine
 
-final class TrackEventViewController: UIViewController {
+final class HomeViewController: UIViewController {
   @IBOutlet private var subscriptionLabel: UILabel!
   private var subscribedCancellable: AnyCancellable?
   private var cancellable: AnyCancellable?
 
-  static func fromStoryboard() -> TrackEventViewController {
+  static func fromStoryboard() -> HomeViewController {
     let storyboard = UIStoryboard(
       name: "Main",
       bundle: nil
     )
     let controller = storyboard.instantiateViewController(
-      withIdentifier: "TrackEventViewController"
-    ) as! TrackEventViewController
+      withIdentifier: "HomeViewController"
+    ) as! HomeViewController
 
     return controller
   }
@@ -35,6 +35,7 @@ final class TrackEventViewController: UIViewController {
   override func viewDidLoad() {
     super.viewDidLoad()
 
+    // subscribe to subscriptionStatus changes
     subscribedCancellable = Superwall.shared.$subscriptionStatus
       .receive(on: DispatchQueue.main)
       .sink { [weak self] status in
@@ -47,12 +48,13 @@ final class TrackEventViewController: UIViewController {
           self?.subscriptionLabel.text = "You do not have an active subscription so the paywall will show when clicking the button."
         }
       }
+
+
     navigationItem.hidesBackButton = true
   }
 
   @IBAction private func logOut() {
-    UserDefaults.standard.setValue(false, forKey: "IsLoggedIn")
-    SuperwallService.reset()
+    Superwall.shared.reset()
     _ = self.navigationController?.popToRootViewController(animated: true)
   }
 
@@ -89,39 +91,4 @@ final class TrackEventViewController: UIViewController {
     self.present(alertController, animated: true)
   }
 
-  // The below function gives an example of how to track an event using Combine publishers:
-  /*
-  func trackEventUsingCombine() {
-    cancellable = Superwall.shared
-      .publisher(forEvent: "MyEvent")
-      .sink { paywallState in
-        switch paywallState {
-        case .presented(let paywallInfo):
-          print("paywall info is", paywallInfo)
-        case .dismissed(let result):
-          switch result.state {
-          case .closed:
-            print("User dismissed the paywall.")
-          case .purchased(productId: let productId):
-            print("Purchased a product with id \(productId), then dismissed.")
-          case .restored:
-            print("Restored purchases, then dismissed.")
-          }
-        case .skipped(let reason):
-          switch reason {
-          case .noRuleMatch:
-            print("The user did not match any rules")
-          case .holdout(let experiment):
-            print("The user is in a holdout group, with experiment id: \(experiment.id), group id: \(experiment.groupId), paywall id: \(experiment.variant.paywallId ?? "")")
-          case .eventNotFound:
-            print("The event wasn't found in a campaign on the dashboard.")
-          case .userIsSubscribed:
-            print("The user is subscribed.")
-          case .error(let error):
-            print("Failed to present paywall. Consider a native paywall fallback", error)
-          }
-        }
-      }
-  }
-  */
 }